Migrate save_geoh5 directive to separate module
